back/requirements.txt
fastapi
uvicorn[standard]
pwnedpasswords-offline
ollama
httpx
pydantic
python-dotenv
numpy
scikit-learn
tensorflow==2.15.0
torch>=2.1.0
transformers>=4.36.0